+++ java-client/src/main/java/io/sphere/internal/request/QueryRequestImpl.java
@@ -14,6 +14,7 @@ public class QueryRequestImpl<T> implements QueryRequest<T>, TestableRequest {
     RequestHolder<QueryResult<T>> requestHolder;
     TypeReference<QueryResult<T>> jsonParserTypeRef;
     private String where = "";
+    private String sort = "";
     private int pageSize = Defaults.pageSize;
     private int page = 0;
 
@@ -27,6 +28,16 @@ public class QueryRequestImpl<T> implements QueryRequest<T>, TestableRequest {
             where = predicate;
         } else {
             where = "";
+        }
+        return this;
+    }
+
+    @Override
+    public QueryRequest<T> sort(String sort) {
+        if(sort != null) {
+            this.sort = sort;
+        } else {
+            this.sort = "";
         }
         return this;
     }
@@ -50,6 +61,9 @@ public class QueryRequestImpl<T> implements QueryRequest<T>, TestableRequest {
     @Override public ListenableFuture<QueryResult<T>> fetchAsync() {
         if(!isNullOrEmpty(where)) {
             requestHolder.addQueryParameter("where",  where);
+        }
+        if(!isNullOrEmpty(sort)) {
+            requestHolder.addQueryParameter("sort",  sort);
         }
         requestHolder.addQueryParameter("limit", Integer.toString(this.pageSize));
         requestHolder.addQueryParameter("offset", Integer.toString(this.page * this.pageSize));